Refactoring obod
